% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convrg_func.R
\name{indexer}
\alias{indexer}
\title{Determine indices for treatment groups}
\usage{
indexer(x)
}
\arguments{
\item{x}{Vector input.}
}
\value{
A list with two named elements that contain start and end indices.
}
\description{
This function determines the indices for the first and last replicates within a vector describing treatment group.
}
\examples{
indexer(c(rep("treatment1",5), rep("treatment2",5)))
}
