% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVRG-package.R
\docType{package}
\name{CNVRG-package}
\alias{CNVRG-package}
\alias{CNVRG}
\title{The 'CNVRG' package.}
\description{
This package implements Dirichlet multinomial modeling of relative abundance data using functionality provided by the 'Stan' software. The purpose of this package is to provide a user friendly way to interface with 'Stan' that is suitable for those new modelling.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2.
}
